requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
DateTime==5.4
exceptiongroup==1.2.0
fastapi==0.104.1
japanize-matplotlib==1.1.3
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
lightning-utilities==0.10.0
matplotlib==3.8.2
Pillow==10.1.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==2.1.2
scikit-learn==1.3.2
shap==0.44.0
streamlit==1.28.2
torch==2.1.1
torchmetrics==1.2.0
torchvision==0.16.1
uvicorn==0.24.0.post1